♻️ refactor: 폴더 구조 내부에서 정한대로 다시 수정
