+++ src/app/quiz/components/Comment/LikeButton.tsx
@@ -6,7 +6,7 @@ import { HTMLAttributes, useState } from 'react';
 
 import { ICON_URL } from '@/common/constants';
 
-import { Text } from '../../../shared';
+import { Text } from '../../../../common/components';
 
 interface LikeButtonProps
   extends Omit<HTMLAttributes<HTMLSpanElement>, 'className'> {
